LayoutTests/accessibility/custom-elements/menuitem-expected.txt

This tests that aria fallback roles work correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS accessibilityController.accessibleElementById("item-1").isEnabled is false
PASS accessibilityController.accessibleElementById("item-1").isExpanded is true
PASS accessibilityController.accessibleElementById("item-1").hasPopup is true
PASS accessibilityController.accessibleElementById("item-2").isEnabled is true
PASS accessibilityController.accessibleElementById("item-2").isExpanded is false
PASS accessibilityController.accessibleElementById("item-2").hasPopup is false
PASS successfullyParsed is true

TEST COMPLETE
